fix: add os-specific implementation for write perm
